Apply same changes to machine-02

+++ project/machine-01/src/main/java/ch/unisg/machine01/core/services/camunda/PublishMachineFillLevelEvent.java
@@ -28,7 +28,7 @@ public class PublishMachineFillLevelEvent {
 
         System.out.println("Event: Publish Machine Fill Level");
 
-        int machineFillLevel = machine.getMachineFillLevel().getValue();
+        double machineFillLevel = machine.getMachineFillLevel().getValue();
 
         HashMap variables = new HashMap();
         variables.put("machineFillLevel", machineFillLevel);
